Convert RSSI to dBm
Instead of returning the raw modem value, convert the RSSI
to dBm. The API of getRSSI() changed to return int16_t.

+++ cicada/commdevices/sim800.cpp
@@ -179,7 +179,7 @@ void Sim800CommDevice::run()
         return;
 
     // When signal strength was requested, send the command to the modem
-    if (_rssi == UINT8_MAX && _stateBooleans & LINE_READ) {
+    if (_rssi == INT16_MAX && _stateBooleans & LINE_READ) {
         _replyState = csq;
         _waitForReply = _okStr;
         sendCommand("AT+CSQ");
